a custom hook for everything on the Calculate/Result pages. Hooray!
Integrated with the Calculate page only.
Integration with the Result page is next
